declare 3 tests with testify
